Fix ExtensionPoint circular dependency
